* fixed sorting culler issue with nested effects * fixed state id assignment bug * fixed state handling of local effects * clean up

+++ Wire/Renderer/GXRenderer/WireGXIndexBuffer.h
@@ -29,10 +29,10 @@ public:
 	void Enable(Renderer* pRenderer);
 	void Disable(Renderer* pRenderer);
 
-	inline TMap<UShort, PdrDisplayList*>& GetDisplayLists();
+	inline TMap<UInt, PdrDisplayList*>& GetDisplayLists();
 
 private:
-	TMap<UShort, PdrDisplayList*> mDisplayLists;
+	TMap<UInt, PdrDisplayList*> mDisplayLists;
 };
 
 #include "WireGXIndexBuffer.inl"
